Move shared reusable helpers out of test class.
